Move FastqReadCounter from 'stats' module to 'fastq_utils' module.
