+++ auto_process_ngs/stats.py
@@ -14,8 +14,6 @@ a run:
 - FastqStatistics: collects and reports stats on FASTQs from an
   Illumina sequencing run
 - FastqStats: container for storing data about a FASTQ file
-- FastqReadCounter: implements various methods for counting reads
-  in FASTQ files
 - collect_fastq_data: collect data from FASTQ file in a FastqStats
   instance
 
@@ -28,12 +26,11 @@ a run:
 import sys
 import os
 import time
-import subprocess
 from multiprocessing import Pool
-import bcftbx.FASTQFile as FASTQFile
 import bcftbx.utils as bcf_utils
 from bcftbx.IlluminaData import IlluminaFastq
 from bcftbx.TabFile import TabFile
+from .fastq_utils import FastqReadCounter
 
 # Initialise logging
 import logging
@@ -482,111 +479,6 @@ class FastqStats:
         Read number extracted from the FASTQ name
         """
         return IlluminaFastq(self.name).read_number
-
-class FastqReadCounter:
-    """
-    Implements various methods for counting reads in FASTQ file
-
-    The methods are:
-
-    - simple: a wrapper for the FASTQFile.nreads() function
-    - fastqiterator: counts reads using FASTQFile.FastqIterator
-    - zcat_wc: runs 'zcat | wc -l' in the shell
-    - reads_per_lane: counts reads by lane using FastqIterator
-
-    """
-    @staticmethod
-    def simple(fastq=None,fp=None):
-        """
-        Return number of reads in a FASTQ file
-
-        Uses the FASTQFile.nreads function to do the counting.
-
-        Arguments:
-          fastq: fastq(.gz) file
-          fp: open file descriptor for fastq file
-
-        Returns:
-          Number of reads
-
-        """
-        return FASTQFile.nreads(fastq=fastq,fp=fp)
-    @staticmethod
-    def fastqiterator(fastq=None,fp=None):
-        """
-        Return number of reads in a FASTQ file
-
-        Uses the FASTQFile.FastqIterator class to do the
-        counting.
-
-        Arguments:
-          fastq: fastq(.gz) file
-          fp: open file descriptor for fastq file
-
-        Returns:
-          Number of reads
-
-        """
-        nreads = 0
-        for r in FASTQFile.FastqIterator(fastq_file=fastq,fp=fp):
-            nreads += 1
-        return nreads
-    @staticmethod
-    def zcat_wc(fastq=None,fp=None):
-        """
-        Return number of reads in a FASTQ file
-
-        Uses a system call to run 'zcat FASTQ | wc -l' to do
-        the counting (or just 'wc -l' if not a gzipped FASTQ).
-
-        Note that this can only operate on fastq files (not
-        on streams provided via the 'fp' argument; this will
-        raise an exception).
-
-        Arguments:
-          fastq: fastq(.gz) file
-          fp: open file descriptor for fastq file
-
-        Returns:
-          Number of reads
-
-        """
-        if fastq is None:
-            raise Exception("zcat_wc: can only operate on a file")
-        if fastq.endswith(".gz"):
-            cmd = "zcat %s | wc -l" % fastq
-        else:
-            cmd = "wc -l %s | cut -d' ' -f1" % fastq
-        output = subprocess.check_output(cmd,shell=True)
-        try:
-            return int(output)/4
-        except Exception,ex:
-            raise Exception("zcat_wc returned: %s" % output)
-    @staticmethod
-    def reads_per_lane(fastq=None,fp=None):
-        """
-        Return counts of reads in each lane of FASTQ file
-
-        Uses the FASTQFile.FastqIterator class to do the
-        counting, with counts split by lane.
-
-        Arguments:
-          fastq: fastq(.gz) file
-          fp: open file descriptor for fastq file
-
-        Returns:
-          Dictionary where keys are lane numbers (as integers)
-            and values are number of reads in that lane.
-
-        """
-        nreads = {}
-        for r in FASTQFile.FastqIterator(fastq_file=fastq,fp=fp):
-            lane = int(r.seqid.flowcell_lane)
-            try:
-                nreads[lane] += 1
-            except KeyError:
-                nreads[lane] = 1
-        return nreads
 
 #######################################################################
 # Functions
