+++ auto_process_ngs/test/test_stats.py
@@ -13,31 +13,9 @@ from bcftbx.IlluminaData import IlluminaFastq
 from bcftbx.IlluminaData import IlluminaData
 from auto_process_ngs.stats import FastqStatistics
 from auto_process_ngs.stats import FastqStats
-from auto_process_ngs.stats import FastqReadCounter
 from auto_process_ngs.stats import collect_fastq_data
 
 # Test data
-fastq_data = """@MISEQ:34:000000000-A7PHP:1:1101:12552:1774 1:N:0:TAAGGCGA
-TTTACAACTAGCTTCTCTTTTTCTT
-+
->AA?131@C1FCGGGG1BFFGF1F3
-@MISEQ:34:000000000-A7PHP:1:1101:16449:1793 1:N:0:TAAGGCGA
-TCCCCAGTCTCAGCCCTACTCCACT
-+
-11>>>11CDFFFBCGGG1AFE11AB
-@MISEQ:34:000000000-A7PHP:1:1101:15171:1796 1:N:0:GCCTTACC
-CCACCACGCCTGGCTAATTTTTTTT
-+
-1>1>>AAAAAFAGGGGBFGGGGGG0
-@MISEQ:34:000000000-A7PHP:1:1101:18777:1797 1:N:0:TAAGNNNA
-CAGCAATATACACTTCACTCTGCAT
-+
-111>>1FBFFFFGGGGCBGEG3A3D
-@MISEQ:34:000000000-A7PHP:1:1101:18622:1812 1:N:0:TAAGGCGA
-GATAAAGACAGAGTCTTAATTAAAC
-+
-11>1>11DFFCFFDGGGB3BF313A
-"""
 fastq_r1_data = """@HISEQ:1:000000000-A2Y1L:1:1101:19264:2433 1:N:0:AGATCGC
 AGATAGCCGA
 +
@@ -1111,117 +1089,6 @@ class TestFastqStats(unittest.TestCase):
         self.assertEqual(fqs.nreads,None)
         self.assertEqual(fqs.fsize,None)
         self.assertEqual(fqs.reads_by_lane,{})
-
-# FastqReadCounter
-class TestFastqReadCounter(unittest.TestCase):
-    def setUp(self):
-        # Temporary working dir (if needed)
-        self.wd = None
-
-    def _make_working_dir(self):
-        if self.wd is None:
-            self.wd = tempfile.mkdtemp(suffix='.test_FastqReadCounter')
-
-    def _make_fastq(self,name,contents):
-        # Create a FASTQ file under the working directory
-        # called "name" and populated with "contents"
-        # Working directory will be created if not already
-        # set up
-        # If "name" ends with ".gz" then "contents" will be
-        # gzipped
-        # Returns the path to the file
-        self._make_working_dir()
-        filen = os.path.join(self.wd,name)
-        if filen.endswith(".gz"):
-            with gzip.GzipFile(filen,'wb') as fp:
-                fp.write(contents)
-        else:
-            with open(filen,'w') as fp:
-                fp.write(contents)
-        return filen
-
-    def tearDown(self):
-        # Remove temporary working dir
-        if self.wd is not None and os.path.isdir(self.wd):
-            shutil.rmtree(self.wd)
-
-    def test_simple(self):
-        readcounter = FastqReadCounter.simple
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),5)
-        fq = self._make_fastq("test_S2_R1_001.fastq",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),12)
-
-    def test_simple_gz(self):
-        readcounter = FastqReadCounter.simple
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq.gz",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),5)
-        fq = self._make_fastq("test_S2_R1_001.fastq.gz",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),12)
-
-    def test_fastqiterator(self):
-        readcounter = FastqReadCounter.fastqiterator
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),5)
-        fq = self._make_fastq("test_S2_R1_001.fastq",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),12)
-
-    def test_fastqiterator_gz(self):
-        readcounter = FastqReadCounter.fastqiterator
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq.gz",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),5)
-        fq = self._make_fastq("test_S2_R1_001.fastq.gz",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),12)
-
-    def test_zcat_wc(self):
-        readcounter = FastqReadCounter.zcat_wc
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),5)
-        fq = self._make_fastq("test_S2_R1_001.fastq",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),12)
-
-    def test_zcat_wc_gz(self):
-        readcounter = FastqReadCounter.zcat_wc
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq.gz",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),5)
-        fq = self._make_fastq("test_S2_R1_001.fastq.gz",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),12)
-
-    def test_reads_per_lane(self):
-        readcounter = FastqReadCounter.reads_per_lane
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),{ 1: 5 })
-        fq = self._make_fastq("test_S2_R1_001.fastq",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),{ 1: 5,
-                                           2: 1,
-                                           3: 2,
-                                           4: 4 })
-
-    def test_reads_per_lane_gz(self):
-        readcounter = FastqReadCounter.reads_per_lane
-        fq = self._make_fastq("test_S1_L001_R1_001.fastq.gz",
-                              fastq_data)
-        self.assertEqual(readcounter(fq),{ 1: 5 })
-        fq = self._make_fastq("test_S2_R1_001.fastq.gz",
-                              fastq_multi_lane_data)
-        self.assertEqual(readcounter(fq),{ 1: 5,
-                                           2: 1,
-                                           3: 2,
-                                           4: 4 })
 
 # collect_fastq_data
 class TestCollectFastqData(unittest.TestCase):
